{"schema":"decodex/commit/1","summary":"Split state runtime row parsers","authority":"XY-1115"}
